redo riding selector
