+++ components/SiteHeader.vue
@@ -138,8 +138,6 @@
 </template>
 
 <script setup>
-import { Disclosure, DisclosureButton, DisclosurePanel, Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/vue'
-import { Bars3Icon, XMarkIcon } from '@heroicons/vue/24/outline'
 import { useAuthStore } from '~/store/auth';
 import SiteHeaderLink from './SiteHeaderLink.vue';
