+++ nuxt.config.ts
@@ -1,26 +1,20 @@
-var neo4jUri = process.env.NEO4J_URI || 'neo4j://localhost:7999';
-var neo4jUsername=process.env.NEO4J_USERNAME || 'neo4j';
-var neo4jPassword=process.env.NEO4J_PASSWORD ||  'password';
-
 export default defineNuxtConfig({
+  app: {
+    head: {
+      script: [
+
+      ]
+    }
+  },
   // Modules
   modules: [
     "@nuxt/content",
     "@nuxtjs/tailwindcss",
-    "nuxt-neo4j",
     '@pinia/nuxt'
   ],
   buildModules:[
     "@nuxtjs/dotenv"
   ],
-  neo4j: {
-    uri: neo4jUri, 
-    auth: {
-      type: 'basic',
-      username: neo4jUsername,
-      password: neo4jPassword,
-    },
-  },
   vite: {
     optimizeDeps: {
       exclude: ['fsevents']
