+++ package.json
@@ -16,32 +16,23 @@
     "@aws-amplify/ui-vue": "^4.2.5",
     "@aws-sdk/client-location": "^3.590.0",
     "@aws/amazon-location-utilities-auth-helper": "^1.0.6",
-    "@headlessui/vue": "^1.7.20",
-    "@heroicons/vue": "^2.1.3",
-    "@neo4j/graphql": "^5.2.0",
-    "@neo4j/graphql-ogm": "^5.2.0",
     "@nuxt/content": "^2.12.1",
     "@pinia/nuxt": "^0.5.1",
-    "@tailwindcss/typography": "^0.5.10",
     "aws-amplify": "^6.3.2",
     "aws-jwt-verify": "^4.0.1",
     "aws-sdk": "^2.1616.0",
     "cytoscape": "^3.29.2",
-    "flowbite": "^2.3.0",
-    "flowbite-vue": "^0.1.4",
     "lodash.debounce": "^4.0.8",
     "neo4j-driver": "^5.18.0",
     "node-fetch": "^3.3.2",
     "nuxt": "^3.10.3",
-    "nuxt-neo4j": "^0.0.4",
-    "path": "^0.12.7",
     "pinia": "^2.1.7",
     "remark-oembed": "^1.2.2",
-    "vue": "^3.4.21",
-    "vue-router": "^4.3.0",
-    "vue-search-select": "^3.2.0"
+    "vue": "^3.4.21"
   },
   "devDependencies": {
+    "@material-tailwind/html": "^2.2.2",
+    "@tailwindcss/typography": "^0.5.10",
     "@nuxtjs/tailwindcss": "^6.11.4",
     "@tailwindcss/forms": "^0.5.7",
     "autoprefixer": "^10.4.18",
